code refactoring phase 2
